requirement.txt
fastapi
uvicorn
supabase
openai
chromadb
pymupdf
python-dotenv
pydantic-settings
python-multipart